fix: 80+ spec failures - MHTML separation, serialization, and API fixes
Key fixes:
- MHTML architecture: Separate class tree from OOXML (Mhtml::Document, etc.)
- Style library path: Fixed STYLES_DIR to point to correct config/styles/
- Style builder: Use property setters for wrapper objects
- add_paragraph: Handle Paragraph objects passed as argument
- DSL autoload: Fix capitalization (Dsl -> DSL)
- Table header: Set properties.header instead of direct attribute
- DocumentRoot: Add render_default: true for body, remove mixed_content
- Line spacing: Use Properties::Spacing for proper XML serialization
- Spec updates: Fix MHTML type expectations, remove obsolete pending tests

Test progress: 282 failures -> 202 failures (80 fixed)

+++ lib/uniword/properties/spacing.rb
@@ -11,7 +11,7 @@ module Uniword
       # Pattern 0: ATTRIBUTES FIRST
       attribute :before, :integer
       attribute :after, :integer
-      attribute :line, :float
+      attribute :line, :integer
       attribute :line_rule, :string
 
       xml do
